Fix some bugs; Finish the game - although still has some adjusts and fix, but is basically finished;

+++ Assets/Script/Misc/PoliceAttackManager.cs
@@ -9,7 +9,7 @@ namespace Assets.Script.Misc
     public class PoliceAttackManager : MonoBehaviour
     {
         public GameObject _policeAttack = null;
-        public int _secondsToWait = 1;
+        public float _secondsToWait = 1;
 
         void Start()
         {
